requirements.txt
whisper
moviepy
transformers
google-generativeai
torch
numpy
scipy
librosa
soundfile
